requirements.txt
alembic==0.9.6
bleach==2.1.2
blinker==1.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
comtypes==1.1.4
cycler==0.10.0
decorator==4.2.1
Django==1.9
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-haystack==2.6.1
django-rest==0.0.1
djangorestframework==3.6.4
dominate==2.3.1
entrypoints==0.2.3
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
html5lib==1.0.1
httpie==0.9.9
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
itchat==1.3.10
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.2.0
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.1
olefile==0.44
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
PyAutoGUI==0.9.36
Pygments==2.0.2
PyMsgBox==1.0.6
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
PyScreeze==0.1.13
python-dateutil==2.6.1
python-editor==1.0.3
PyTweening==1.0.3
pytz==2017.2
pywifi==1.1.9
pywin32==221
pywinpty==0.5.1
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
selenium==3.10.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.0
terminado==0.8.1
testpath==0.3.1
threadpool==1.3.2
tornado==4.5.3
traitlets==4.3.2
virtualenv==15.1.0
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
WTForms==2.1
xlrd==1.1.0
